global search done without pagination

+++ SppotiBusiness/src/main/java/com/fr/service/GlobalSearchService.java
@@ -14,9 +14,9 @@ public interface GlobalSearchService extends AbstractControllerService
 	GlobalSearchResultDTO findAllUsersFromCriteria(String query, String sexe, Integer ageMax, Integer ageMin,
 												   Integer page);
 	
-	GlobalSearchResultDTO findAllTeamFromCriteria(String query, Integer sport, Integer page);
+	GlobalSearchResultDTO findAllTeamFromCriteria(String query, Long sport, Integer page);
 	
-	GlobalSearchResultDTO findAllSppotisFromCriteria(String query, Integer sport, String startDate, Integer page);
+	GlobalSearchResultDTO findAllSppotisFromCriteria(String query, Long sport, String startDate, Integer page);
 	
 	GlobalSearchResultDTO findAllWithoutCriteria(String query, Integer page);
 }
