+++ SppotiRestApi/src/main/java/com/fr/api/find/GlobalSearchController.java
@@ -32,7 +32,7 @@ public class GlobalSearchController
 	public ResponseEntity<GlobalSearchResultDTO> findData(@RequestParam final int type, @RequestParam final String sex,
 														  @RequestParam("age_max") final Integer ageMax,
 														  @RequestParam("age_min") final Integer ageMin,
-														  @RequestParam final Integer sport,
+														  @RequestParam final Long sport,
 														  @RequestParam("start_date") final String startDate,
 														  @RequestParam final String query,
 														  @RequestParam final int page)
